+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ logging.path=/root/hbase-api/log
 spring.output.ansi.enabled=always
 
 # \u670D\u52A1\u5668\u5BF9\u5916\u5F00\u653E\u7AEF\u53E3
-server.port=8008
+server.port=8009
 
 spring.main.web-application-type=REACTIVE
 
@@ -16,6 +16,5 @@ tensor.hbase-api.mq.ring-buffer-size=1024*1024
 spring.application.name=hbase-api
 
 # nacos
-nacos.config.server-addr=120.24.90.180:8848
+nacos.client.ip=127.0.0.1
 nacos.discovery.cluster-name=tensor
-nacos.discovery.server-addr=120.24.90.180:8848
